requirements.txt
# Core Testing Framework
selenium==4.16.0
pytest==7.4.3
pytest-xdist==3.5.0  # Parallel execution
pytest-rerunfailures==13.0  # Retry failed tests
pytest-timeout==2.2.0  # Test timeouts
pytest-html==4.1.1  # HTML reports

# Web Drivers
webdriver-manager==4.0.1  # Auto-manage browser drivers

# Reporting
allure-pytest==2.13.2  # Allure reports
pytest-metadata==3.0.0

# Data Handling
pyyaml==6.0.1  # Config files
python-dotenv==1.0.0  # Environment variables
faker==22.0.0  # Test data generation
jsonschema==4.20.0  # JSON validation

# API Testing
requests==2.31.0
urllib3==2.1.0

# Utilities
pillow==10.1.0  # Screenshot handling
opencv-python==4.8.1.78  # Image comparison
python-dateutil==2.8.2
loguru==0.7.2  # Enhanced logging

# Performance & Monitoring
psutil==5.9.7  # System monitoring
memory-profiler==0.61.0

# Code Quality
black==23.12.1  # Code formatting
pylint==3.0.3  # Linting
mypy==1.7.1  # Type checking
pytest-cov==4.1.0  # Code coverage

# CI/CD
tox==4.11.4  # Testing automation

# Database (for data-driven testing)
sqlalchemy==2.0.23

# Additional
colorama==0.4.6  # Colored terminal output
tabulate==0.9.0  # Pretty tables
